+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Version 0.9.5 (not released yet)
+- Add Deep copy for LoginInformation (**FEATURE**)
 
 ## Version 0.9.4 (released 2020-01-18)
 - Add ReplaceLoginInformationSecret to CommonSecretsContainer (**FEATURE**)