+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Version 0.9.5 (not released yet)
 - Add Deep copy for LoginInformation (**FEATURE**)
 - Add Deep copy for FileEntry (**FEATURE**)
+- Add Deep copy for Note (**FEATURE**)
 
 ## Version 0.9.4 (released 2020-01-18)
 - Add ReplaceLoginInformationSecret to CommonSecretsContainer (**FEATURE**)