+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Add Deep copy for LoginInformation (**FEATURE**)
 - Add Deep copy for LoginInformationSecret (**FEATURE**)
 - Add Deep copy for FileEntry (**FEATURE**)
+- Add Deep copy for FileEntrySecret (**FEATURE**)
 - Add Deep copy for Note (**FEATURE**)
 - Add Deep copy for NoteSecret (**FEATURE**)
 - Add Deep copy for SymmetricKeyAlgorithm (**FEATURE**)